+++ modules/App/assets/js/app/utils.js
@@ -111,7 +111,19 @@ let stripTags = function(input, allowed) {
     }
 };
 
+let base64encode = function(str) {
+    return btoa(encodeURIComponent(str).replace(/%([0-9A-F]{2})/g, (match, p1) =>  String.fromCharCode('0x' + p1)));
+};
+
+let base64decode = function(str) {
+    return decodeURIComponent(atob(str).split('').map((c) => {
+        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2)
+    }).join(''));
+};
+
 export default {
+    base64encode,
+    base64decode,
     copyText,
     formatSize,
     formatDuration,