+++ modules/Content/views/collection/items.php
@@ -349,7 +349,7 @@
 
                             if (searchParams.has('state')) {
                                 try {
-                                    const q = JSON.parse(atob(searchParams.get('state')));
+                                    const q = JSON.parse(App.utils.base64decode(searchParams.get('state')));
                                     if (q.sort) this.sort = q.sort;
                                     if (q.page) this.page = q.page;
                                     if (q.limit) this.limit = (parseInt(q.limit) || this.limit);
@@ -458,7 +458,7 @@
 
                             window.history.pushState(
                                 null, null,
-                                App.route(['/content/collection/items/', this.model.name, '?state=', btoa(JSON.stringify({
+                                App.route(['/content/collection/items/', this.model.name, '?state=', App.utils.base64encode(JSON.stringify({
                                     page: this.page || null,
                                     filter: this.filter || null,
                                     sort: this.sort || null,