+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add additional security check editing php files in finder
 - Fix base64 encoding data not in latin char range
+- Update vendor libs
 
 ## 2.3.9 (2023-02-13)